refactor: move swarm persister to its own package
